requirements.txt
django # Web framework
psycopg2 # PostgreSQL driver
djangorestframework # Token auth